Constructing New Nichewebs
THIS CODE IS GROSS.  I should wrap it all up in an if loop to destroy
excessive for loops.  That way you only calculate this stuff for species
with life stages.

Anyhow, I just proposed a bunch of ways to calculate new lifestages. Will
explain better in the methods section, because this looks like trash. It's
just really ugly code and I don't think anyone can understand my comments
even.

Signed-off-by: Stephanie Bland <stephie.bland@gmail.com>