Linear Regression (Niche Value & Body Mass)
Added T2 (second trophic level metric) as output for TrophicLevels
function.
Added a linear regression function to find a relationship between niche
value and individual body mass. This can be used to estimate the niche
values for new fish life stages.
Bunch of test code to find best way to estimate linear regression.
Still not
done with this. I would prefer to look into it more, because current
method is ok for estimating niche levels of non-plant and non-fish
species, but would not extrapolate well for fish, and there aren’t
enough
fish species in each web to just run a lin regression on them alone.