Cleaning up the Web Discarding Methods
Sorry, I know it’s a giant pain to have unnecessary cosmetic changes to
code, but some of these lines are redundant and I want my methods
section to have a clean correspondence to the code itself.

The first change is a redundancy in the loop. It might be an artifact
of C++, but since we seem married to proprietary software, I think it’s
safe to eliminate it.