Find species with no autotrophs in food chain
Clean up the method for detecting species that don't get energy from
autotrophs.

I’m sorry, but I find it extremely difficult to keep track of 3 lists,
and I feel uncomfortable with changing nicheweb on every iteration. My
method is a *bit* more computationally costly, since the length of the
list increases every iteration, but I think it’s worth it for clarity.

If you wanted to be *really* slick, you could take advantage of this
process to simultaneously calculate trophic level. If any values of
T1==0, it would be an ineligible web, so it can be discarded. But it’s
probably easiest to avoid prematurely explaining T1 in the methods.