Delete abort_sim
I originally built in a limiting condition that prevented “dead” webs
from being constructed. So if fish didn’t survive the unfitted scenario
I wouldn’t even bother with other scenarios.

I eliminated it because it’s too difficult to try to extend it to life
stages being treated as separate nodes, and it only makes sense to
apply it to all 3 experiments (because you can’t compare simulations of
webs that are guaranteed not to die in one experiment with simulations
of them in another loosely controlled experiment without that abort_sim
criterion.)