Failed Attempt to Fix num of years
Working under the assumption that species will eventually stabilize, we
can just run each phase until it’s relatively stable, deterministic,
time series. I tried to do that here, but ran into problems:

1. During the linked stages phase we should really be comparing the
biomass at the end of the year to the biomass at the end of the
previous year (because leslie matrix shifts everything around).

2. Some species never seem to reach extinction - I don’t know how to
stop this because it’s deep down in the ODE (search for ext_thresh, and
you’ll find it).

3. Don’t know what tolerance for stability we should use.

4. Stability could happen over several years - so 5 goes to 1 goes to 5
again (like a cyclic permutation from algebra) so it's cyclic, but the
period is a multiple of several years. If I want to fix this, I'd have
to check the current B0 against any historical ones.