FIX LOGICAL SWITCHES in Parameters
Really embarrassing mistake - I set a bunch of switches that you can
assign with true/false or some other values, and the problem is that
matlab doesn’t distinguish between a logical and a numeric.  So I need
to treat true=1 and false=0 as switches that I already used (I mean you
can circumvent this problem with careful isnumeric statements, but I
think it’s too risky and easy to mess up.