Explanation of Error
I found a rare error that seems to occur in less than 1% of all runs.
Simulations that trigger the error include:

RunCluster(0,253,253,2,1) - this will trigger it for year 8 (i=8)
RunCluster(0,271,271,0,1) - this will trigger it for year 5 (i=5)

(Both cases are triggered on Exper=2)

The following is speculation:

Background on how ode45 works: (I think this is how it works at least)
this function wants to minimize the number of time steps to graph the
function, so it predicts where the function will be at time t+delta,
and if |f(t)-f(t+delta)|<epsilon (meaning the error is small enough),
than it will use delta as the time step. Of course, we want the f(t)
for integer values of t, so ODE then uses the points along the straight
line connecting two of it’s time steps to calculate it for our desired
locations.

I believe this error occurs when species go extinct in the very first
time step of the ODE. The ODE has an events function that will pause
the ODE whenever a species’ biomass goes below the extinction
threshold. When it’s paused, the ODE goes back to the last integer time
step and forces the doomed species to an early extinction. The problem
is when this happens in the very first time step, say after 0.5 days. I
believe it just doesn’t pause when that happens, because you end up
just skipping that extinction and lumping it with the next time step
instead. I’m not really sure why it gets skipped.

The original model doesn’t seem to have this problem, which is
consistent with my beliefs: our model restarts the ODE every year while
the original one just runs a continuous ODE for the entire simulation
decade.

You can *forcibly* reproduce the error in the original model (first
commit, webdriver) if you use the exact same initial conditions for the
ODE. So pause it right before it enters the ODE again on the doomed
year and use those initial conditions to run the dynamic_fn in the
original model.