Debugging - Find Root Of Error
The root of the problem is right here in dynamic_fn, the first
extinction event happens immediately, so it’s skipped (reason unknown).
The next extinction event that happens that year now has two events to
deal with, and it doesn’t deal with it in the same vector, rather it
doubles the vector size. I don’t understand why it does this, but I
know how to fix it.