Correct Time Stamps
Confirmed! This Works! We can double check by looking at the time
the simulation took to run against the time the file was saved
locally on your laptop/mac:

sed '/r/p' progress*
ls -l progress*

Of course, it could be completely wrong for one or two files if
your computer was off at the time, but most of the time it should
be accurate.