Limit Number of Computational Threads
https://www.mathworks.com/help/matlab/ref/maxnumcompthreads.html

See email for more details. This is what Ross Dickson (ACENET Research
Consultant) said:

From 'ps' it looks very much like your compiled Matlab code is running
in multiple threads, but you've only reserved a single core for each
job in Grid Engine. Did you intend to run the code multithreaded? That
is, are you using any parallel constructs like parfor or parfeval? If
so, you should be submitting your jobs with "-pe openmp N", with some
suitable N. If on the other hand Matlab is quietly running in parallel
unasked, then we need to figure out how to make it NOT do that.

Yeah, I know it's compiled Matlab, etc. Space (memory) is not an issue;
it's CPUs that are being oversubscribed. Since your job script contains
no "-pe" clause, it's requesting the default one CPU for each job. But
the compiled MATLAB code is not respecting that, and trying to use two
or more CPUs. Now that I know to look for it, I can see it was also
doing that in (at least some of) the jobs from Placentia that you
recorded in that maxvmem log. 'qacct -j 9830193', for example, shows
wallclock of 41436 seconds and cpu of 45964 seconds. The second can
only be larger than the first if it's using two or more CPUs sometimes.

"LASTN = maxNumCompThreads(N) sets the maximum number of computational
threads to N..." sounds exactly like what I was looking for. So, if you
insert a line "maxNumCompThreads(1)" and recompile the code, I think it
should do The Right Thing.

Alternatively, you could set maxNumCompThreads(2) and then submit the
job with "#$ -pe openmp 2", or 3 or whatever, but we probably don't
need to introduce that complexity into your workflow. I suspect
maxNumCompThreads(1) will be optimal.