(MAJOR FIX) Corrected ODE reproductive effort
So the reproductive effort should work somewhat parallel to the regular
metabolic growth. As in the ODE is probably correctly formatted for
normal biomass growth, and I can model reproductive effort off of that.

Essentially, reproductive effort can be treated identically to reg
growth, and so it should follow the same steps it would use.