ODE runaway growth & mislabeled fish as autotrophs
TWO PROBLEMS: 1) ODE wrongly allows runaway growth and 2) mislabeled
fish as autotroph.

1) okay obviously what’s happening is you have a runaway process with
incorrect ODE settings - the ode function is complaining it can’t make
time step delta small enough that the jump in the function is less than
epsilon. This occurs in year 120 of RunCluster(0,290,290,1,0).

So I checked out the plots, and what do you know, they increase
linearly on a *LOG* plot. As in, they’re really increasing
exponentially…. That’s really really bad. Means something is super
messed up with the ODEs that I’m using. It looks like two of the fish
species feed on each other, and are somehow able to contribute more
than 100% of their energy into each other’s growth. Or maybe it’s
reproduction that’s messy. How should I know?

2) My other problem is that one of my fish life stages identifies as an
autotroph and I don’t want to bring gender politics into this debacle.

Clearly it's because they weren't given a proper diet.

But on another level: does it even really matter if we end up with
autotrophic fish??? I mean, A) it would really mess up the analysis if
we aren't careful (because that node might be counted twice or
discarded depending on how things are defined), and B) what if it gets
assigned an autotrophic intrinsic growth rate???? OH NOES!!!

the other thing to consider is that we might end up with "unintentional
autotrophic fish" - as in fish life stages with extinct prey species.
but at least those are guaranteed not to have an intrinsic growth rate
and I don't really care about the nitpicky details of "oh no how could
it survive a year with no food" just like assume it's given parental
care for that year or something dumb like that i dont care it's not my
problem these are fictional fish i can make up whatever stories i want
for them.