2017Nov13_2 TEST Simplified Leslie Matrix
Nodes are still connected, but in a less complicated fashion. This is
to see if the weird bug I have in that other branch will actually cause
issues.

It’s possible that the weird bug is the root cause of the more
stable dynamics in the linked experiments - if my leslie matrices
provide a boost to the fish population, it’s kind of “cheating” in a
sense, because my linked experiments would make up for any
underperforming nodes using that artificial and unrealistic boost.