First Draft-probably defective
WILL IT WORK???????

Problem: [Ticket#015341] over quota at Mahone, jobs jammed

Ross at ACENET says:
Let me see how well I understand the issues: You're running a boatload
of simulations in compiled MATLAB, something like 5000 of them, right?
And each simulation, at each of 500 timesteps, writes a file to disk
containing a dataframe? So something around 5000*500 = 2,500,000 files
expected?

The redundant variables in the dataframe don't worry me --- the total
volume of data you've stored so far in 200,000 files at Mahone is under
5GB. That might scale up to 75GB or so if we had all 2.5M files, which
is no big deal. Yeah, it could be made smaller, but you are quite
correct that having the redundancy is better from a clarity and
simplicity point-of-view, so don't change that.

The problem is the number of files, and the risk that poses to
filesystem performance.

If that's not feasible, my next idea is to bundle up all 500 files from
one simulation into a single file after MATLAB is finished, and then
delete the 500. I'd use tar for this
(https://docs.computecanada.ca/wiki/Archiving_and_compressing_files) by
adding a few lines to the job script: